# API/docs/postman/swagger.yaml
...
swagger: "2.0"
info:
  description: "Panalpina Swagger API server."
  version: "1.0.0"
  title: "Panalpina API"
  termsOfService: "http://panalpinaAPI.com/terms/"
  contact:
    email: "panalpinaTeam@panalpinaAPI.com"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "panalpinaAPI.com"
basePath: "/v1"
schemes:
- "https"
- "http"
paths:
  /shipment/{trackingNumber}:
    get:
      summary: "Get shipment type based on tracking number."
      description: ""
      operationId: "getFlight"
      produces:
        - "application/json"
      parameters:
        - name: "trackingNumber"
          in: "path"
          description: "tracking shipment number."
          required: true
          type: "number"
      responses:
        200:
          description: "Success response with shipment type details."
          schema:
            $ref: "#/definitions/ShipmentResponse"
        404:
          description: "404 Failure response for non existing shipment."
          schema:
            type: "object"
            properties:
              code:
                type: "number"
              message:
                type: "string"
        500:
          description: "500 For internal server errors."
          schema:
            type: "object"
            properties:
              code:
                type: "number"
              message:
                type: "string"
      x-swagger-router-controller: "Shipment"
...
  /freight/{shipmentNumber}:
    get:
      summary: "Get freight shipment details."
      description: ""
      operationId: "getFreightShipment"
      produces:
        - "application/json"
      parameters:
        - name: "shipmentNumber"
          in: "path"
          description: "Freight shipment number."
          required: true
          type: "number"
      responses:
        200:
          description: "Success response with shipment details."
          schema:
            $ref: "#/definitions/FreightShipmentResponse"
        404:
          description: "404 Failure response for non existing shipment."
          schema:
            type: "object"
            properties:
              code:  # diagnostic error codes
                type: "number"
              message:
                type: "string"
        500:
          description: "500 For internal server errors."
          schema:
            type: "object"
            properties:
              code:
                type: "number"
              message:
                type: "string"
      x-swagger-router-controller: "Shipment"
    post:
      summary: "Update freight shipment details."
      description: ""
      operationId: "updateFreightShipment"
      produces:
        - "application/json"
      parameters:
        - name: "shipmentNumber"
          in: "path"
          description: "Flight shipment number."
          required: true
          type: "number"
        - in: "body"
          name: "body"
          description: "update shipment details"
          required: true
          schema:
            $ref: "#/definitions/FreightShipmentResponse"
      responses:
        200:
          description: "Success updates on a Freight shipment"
        404:
          description: "404 Failure response for non existing shipment."
          schema:
            type: "object"
            properties:
              code:
                type: "number"
              message:
                type: "string"
        500:
          description: "500 For internal server errors."
          schema:
            type: "object"
            properties:
              code:
                type: "number"
              message:
                type: "string"
      x-swagger-router-controller: "Shipment"
definitions:
  ShipmentResponse:
    type: "object"
    properties:
      type:
        type: "string"
...
  FreightShipmentResponse:
    type: "object"
    properties:
      data:
        type: "object"
        properties:
          "paintainerHBL": 
            type: "string"
          "status": 
            type: "string"
          "customerReference": 
            type: "string"
          "lastUpdate":
            type: "string"
          "issueDate": 
            type: "string"                  
          "issuePlace": 
            type: "string"
          "receivedDate": 
            type: "string"               
          "releasedDate": 
            type: "string"               
          "shippingLine": 
            type: "string"               
          "placeOfReceipt": 
            type: "string"            
          "placeOfDelivery": 
            type: "string"            
          "vessel": 
            type: "string"                     
          "carrier": 
            type: "string"
          "portOfLoading": 
            type: "string"              
          "portOfDischarge": 
            type: "string"            
          "voyageNumber": 
            type: "string"               
          "incoterms": 
            type: "string"
          "etd": 
            type: "string"                        
          "eta": 
            type: "string"                        
          "totalPieces": 
            type: "string"                
          "weight": 
            type: "string"                     
          "volume": 
            type: "string"                     
          "containerNumber": 
            type: "string"
          "size":
            type: "string"
          "sealNumber": 
            type: "string"
      
          "events":
              type: "array"
              items:
                type: "object"
                properties:
                  date:
                    type: "string"
                  event:
                    type: "string"
                  
          "packageGeoInfo":  
            type: "array"
            items:
              type: "object"
              properties:
                "vesselCurrentPosition":
                 type: "object"
                 properties:
                  "mmsi":
                    type: "string"
                  "lat":
                    type: "string"
                  "lon": 
                    type: "string"
                  "timestamp":
                    type: "string"
                                 
                "vesselVoyageInfo":
                  type: "object"
                  properties:
                    "route": 
                      type: "array"
                      items:
                        type: "object"  
                        properties:                  
                          "lat": 
                            type: "string"
                          "lon": 
                            type: "string"
                  
                    "distanceTravelled": 
                      type: "string"
                    "distanceToGo": 
                      type: "string"
                            
                "vesselHistoricPosition": 
                  type: "array"
                  items:
                     type: "object"                  
                     properties:
                       "lat": 
                         type: "string"
                       "lon": 
                         type: "string"
                       "timestamp": 
                         type: "string"
...